refactor(tests): remove deprecated values
