+++ packages/server/test/_testHelpers.ts
@@ -2,23 +2,27 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import AbortController from 'abort-controller';
 import fetch from 'node-fetch';
-import {
-  createWSClient,
-  WebSocketClientOptions,
-  TRPCWebSocketClient,
-} from '../../client/src/links/wsLink';
 import ws from 'ws';
 import { createTRPCClient, CreateTRPCClientOptions } from '../../client/src';
-import { AnyRouter, CreateHttpHandlerOptions } from '../src';
-import { createHttpServer } from '../src';
-import { applyWSSHandler, WSSHandlerOptions } from '../src/ws';
+import { applyWSSHandler } from '../src/adapters/ws';
+import { WSSHandlerOptions } from '../src/ws';
+import {
+  createWSClient,
+  TRPCWebSocketClient,
+  WebSocketClientOptions,
+} from '../../client/src/links/wsLink';
+import { AnyRouter } from '../src';
+import {
+  CreateHTTPHandlerOptions,
+  createHTTPServer,
+} from '../src/adapters/standalone';
 
 (global as any).fetch = fetch;
 (global as any).AbortController = AbortController;
 export function routerToServerAndClient<TRouter extends AnyRouter>(
   router: TRouter,
   opts?: {
-    server?: Partial<CreateHttpHandlerOptions<TRouter>>;
+    server?: Partial<CreateHTTPHandlerOptions<TRouter>>;
     wssServer?: Partial<WSSHandlerOptions<TRouter>>;
     wsClient?: Partial<WebSocketClientOptions>;
     client?:
@@ -35,7 +39,7 @@ export function routerToServerAndClient<TRouter extends AnyRouter>(
   },
 ) {
   // http
-  const httpServer = createHttpServer({
+  const httpServer = createHTTPServer({
     router,
     createContext: ({ req, res }) => ({ req, res }),
     ...(opts?.server ?? {
