+++ packages/server/test/errors.test.ts
@@ -9,6 +9,7 @@ import { OnErrorFunction } from '../src/internals/OnErrorFunction';
 import { getMessageFromUnkownError } from '../src/internals/errors';
 import { TRPCError } from '../src/TRPCError';
 import { routerToServerAndClient, waitError } from './_testHelpers';
+import { CreateHTTPContextOptions } from '../adapters/standalone';
 
 test('basic', async () => {
   class MyError extends Error {
@@ -128,13 +129,13 @@ describe('formatError()', () => {
       trpc
         .router()
         .formatError(({ error, shape }) => {
-          if (error.originalError instanceof ZodError) {
+          if (error.cause instanceof ZodError) {
             return {
               ...shape,
               data: {
                 ...shape.data,
                 type: 'zod' as const,
-                errors: error.originalError.errors,
+                errors: error.cause.errors,
               },
             };
           }
@@ -206,7 +207,7 @@ Object {
     expect(onError).toHaveBeenCalledTimes(1);
     const serverError = onError.mock.calls[0][0].error;
 
-    expect(serverError.originalError).toBeInstanceOf(ZodError);
+    expect(serverError.cause).toBeInstanceOf(ZodError);
 
     close();
   });
@@ -232,7 +233,7 @@ Object {
       trpc
         .router()
         .formatError(({ error, shape }) => {
-          if (!(error.originalError instanceof ZodError)) {
+          if (!(error.cause instanceof ZodError)) {
             return shape;
           }
           return {
@@ -268,7 +269,7 @@ Object {
     const onError = jest.fn();
     const { close, httpUrl } = routerToServerAndClient(
       trpc
-        .router<trpc.CreateHttpContextOptions>()
+        .router<CreateHTTPContextOptions>()
         .middleware(({ ctx }) => {
           ctx.res.statusCode = TEAPOT_ERROR_CODE;
           throw new Error('Some error');
@@ -382,7 +383,7 @@ test('retain stack trace', async () => {
   const serverOnErrorOpts = onError.mock.calls[0][0];
   const serverError = serverOnErrorOpts.error;
   expect(serverError).toBeInstanceOf(TRPCError);
-  expect(serverError.originalError).toBeInstanceOf(CustomError);
+  expect(serverError.cause).toBeInstanceOf(CustomError);
 
   expect(serverError.stack).not.toContain('getErrorFromUnknown');
   const stackParts = serverError.stack!.split('\n');
